refactor: for consistency
